Moved the architecture dependent processing below  "arch_i686_unknown_linux_gnu” and “arch_arm_none_eabi” packages.
