+++ kernel/arch/arm-none-eabi/Cargo.toml
@@ -1,3 +1,6 @@
 [package]
 name = "arch_arm_none_eabi"
 version = "0.0.1"
+
+[dependencies]
+ructiss_core = { path = "../../common/core" }
