+++ kernel/arch/i686-unknown-linux-gnu/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "arch"
+name = "arch_i686_unknown_linux_gnu"
 version = "0.0.1"
 
 [dependencies]
